+++ hpc/nvidia/cuda/adapter/include/Openacc/cupti_openacc.h
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
+/* Copyright (C) 2026 Rong Tao */
+#ifndef __CUDA_ADAPTER__CUPTI_OPENACC_H_
+#define __CUDA_ADAPTER__CUPTI_OPENACC_H_ 1
+
+#ifdef _CUPTI_OPENACC_H_
+#error "CudaAdapter not allow include origin CUDA Openacc/cupti_openacc.h"
+#endif
+
+#include "wrapper_defs.h"
+
+#endif