+++ hpc/nvidia/cuda/adapter/include/Openmp/cupti_openmp.h
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
+/* Copyright (C) 2026 Rong Tao */
+#ifndef __CUDA_ADAPTER__CUPTI_OPENMP_H_
+#define __CUDA_ADAPTER__CUPTI_OPENMP_H_ 1
+
+#ifdef _CUPTI_OPENMP_H_
+#error "CudaAdapter not allow include origin CUDA Openmp/cupti_openmp.h"
+#endif
+
+#include "wrapper_defs.h"
+
+#endif